rename Permalinkable
