+++ app/models/communication/website/agenda/event.rb
@@ -45,6 +45,7 @@
 class Communication::Website::Agenda::Event < ApplicationRecord
   include AsDirectObject
   include Contentful
+  include Permalinkable
   include Sanitizable
   include Sluggable
   include WithAccessibility
@@ -53,7 +54,6 @@ class Communication::Website::Agenda::Event < ApplicationRecord
   include WithDuplication
   include WithFeaturedImage
   include WithMenuItemTarget
-  include WithPermalink
   include WithTime
   include WithTranslations
   include WithTree
