+++ app/models/communication/website/page.rb
@@ -54,6 +54,7 @@ class Communication::Website::Page < ApplicationRecord
 
   include AsDirectObject
   include Contentful
+  include Permalinkable
   include Sanitizable
   include Sluggable # We override slug_unavailable? method (and set_slug and skip_slug_validation? in Page::Home)
   include WithAccessibility
@@ -65,7 +66,6 @@ class Communication::Website::Page < ApplicationRecord
   include WithType # WithType can set default publication status, so must be included before WithPublication
   include WithPublication
   include WithPosition # Scope :ordered must override WithPublication
-  include WithPermalink
   include WithTranslations
   include WithTree
   include WithPath # Must be included after Sluggable. WithPath overwrites the git_path method defined in WithWebsites
