+++ app/models/communication/website/portfolio/category.rb
@@ -39,13 +39,13 @@
 class Communication::Website::Portfolio::Category < ApplicationRecord
   include AsDirectObject
   include Contentful
+  include Permalinkable
   include Sanitizable
   include Sluggable # We override slug_unavailable? method
   include Pathable # Included after Sluggable to make sure slug is correct before anything
   include WithBlobs
   include WithFeaturedImage
   include WithMenuItemTarget
-  include WithPermalink
   include WithPosition
   include WithTranslations
   include WithTree
