+++ app/models/concerns/permalinkable.rb
@@ -1,8 +1,8 @@
-module WithPermalink
+module Permalinkable
   extend ActiveSupport::Concern
 
   included do
-    include WithHugo
+    include Staticable
 
     has_many  :permalinks,
               class_name: "Communication::Website::Permalink",
