+++ app/models/education/program.rb
@@ -66,6 +66,7 @@
 class Education::Program < ApplicationRecord
   include AsIndirectObject
   include Contentful
+  include Permalinkable
   include Sanitizable
   include Sluggable
   include Pathable # Included after Sluggable to make sure slug is correct before anything
@@ -79,7 +80,6 @@ class Education::Program < ApplicationRecord
   include WithInheritance
   include WithLocations
   include WithMenuItemTarget
-  include WithPermalink
   include WithPosition
   include WithSchools
   include WithTeam
