+++ app/models/research/journal/volume.rb
@@ -38,12 +38,12 @@
 #
 class Research::Journal::Volume < ApplicationRecord
   include AsIndirectObject
+  include Permalinkable
   include Sanitizable
   include Sluggable
   include WithBlobs
   include WithFeaturedImage
   include WithGitFiles
-  include WithPermalink
   include WithPublication
   include WithTranslations
   include WithUniversity
